[Journal] Updating create RR draft files to match new format
