+++ src/researchhub/urls.py
@@ -202,6 +202,11 @@ urlpatterns = [
         + (settings.HEALTH_CHECK_TOKEN + "/" if settings.HEALTH_CHECK_TOKEN else ""),
         include("health_check.urls"),
     ),
+    path(
+        "api/researchhubpost/create_registered_report_draft/",
+        researchhub_document_views.RegisteredReportDraftView.as_view(),
+        name="registered-report-draft",
+    ),
     re_path(r"^api/", include(router.urls)),
     # Nested routes for list items
     path(
