+++ src/researchhub_document/views/__init__.py
@@ -1,3 +1,4 @@
 # flake8: noqa
+from .registered_report_draft_view import RegisteredReportDraftView
 from .researchhub_post_views import ResearchhubPostViewSet
 from .researchhub_unified_document_views import ResearchhubUnifiedDocumentViewSet
